Move experiement ingestion into experiment endpoint
